Register Requests, use of both databases and frontend enhancements.
This app allows only students or professor, they can request a register application, if their application data through the app matches the documents of an already employed/registered professor/student of the University, their account request for the app will be approved and they can have access to the app, if not the request will be denied.

MySQL and MongoDB relationships.